+++ CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
+Unreleased
+----------
+
+* Fixed / improved `PageLanguageOverlayCommandController`
+* Updated AppVeyor configuration
+* Cleaned up
+
 4.4.0
 -----